bitcoin-price-pipeline/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-postgres
psycopg2-binary
pandas

# Environment variables management
python-dotenv>=1.0.0